# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Buckhill Petshop Api Documentation",
        "description": "Buckhill Petshop Api Documentation",
        "contact": {
            "name": "Kelvin Mukotso",
            "email": "kelvinmukotso@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api/v1"
        }
    ],
    "paths": {
...
        "/main/blog": {
            "get": {
                "tags": [
                    "Main Page"
                ],
                "summary": "List all blogs",
                "description": " view all blogs",
                "operationId": "11b13303f385fabd5ad00463e0275800",
                "responses": {
                    "200": {
                        "description": "Successfull (Ok)"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Page not found"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
        "/main/blog/{blog_uuid}": {
            "get": {
                "tags": [
                    "Main Page"
                ],
                "summary": "view a single blog",
                "description": "view a blog",
                "operationId": "8ffc71d837feae92b3a973192a070a28",
                "parameters": [
                    {
                        "name": "blog_uuid",
                        "in": "path",
                        "description": "blog id",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfull (Ok)"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Page not found"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
        "/main/promotions": {
            "get": {
                "tags": [
                    "Main Page"
                ],
                "summary": "List all promotions",
                "description": " view all promotions",
                "operationId": "ba7d3e99826451d580f7b525b72a8c94",
                "responses": {
                    "200": {
                        "description": "Successfull (Ok)"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Page not found"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
...
    }
}